% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcftools.smry.funcs.R
\name{vcftools.idepth}
\alias{vcftools.idepth}
\alias{vcftools.siteDepth}
\alias{vcftools.genoDepth}
\alias{vcftools.imiss}
\alias{vcftools.siteMiss}
\alias{vcftools.freqBurden}
\alias{vcftools.freq2}
\alias{vcftools.singletons}
\alias{vcftools.012}
\alias{vcftools.het}
\alias{vcftools.siteQual}
\title{Title vcftools Summary Functions}
\usage{
vcftools.idepth(vcf.fn, res.fn)

vcftools.siteDepth(vcf.fn, res.fn)

vcftools.genoDepth(vcf.fn, res.fn)

vcftools.imiss(vcf.fn, res.fn)

vcftools.siteMiss(vcf.fn, res.fn)

vcftools.freqBurden(vcf.fn, res.fn)

vcftools.freq2(vcf.fn, res.fn)

vcftools.singletons(vcf.fn, res.fn)

vcftools.012(vcf.fn, res.fn)

vcftools.het(vcf.fn, res.fn)

vcftools.siteQual(vcf.fn, res.fn)
}
\arguments{
\item{vcf.fn}{character string with the relative path and name of the vcf file to be filtered}

\item{res.fn}{character string with the relative path and name of the output vcf file}
}
\value{

}
\description{
These functions filter calculate summary info for individuals and
    loci. Input and results file names should include the path, but no
    extension.
}
