% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcftools.loc.filters.R
\name{vcftools.minDP}
\alias{vcftools.minDP}
\alias{vcftools.minQ}
\alias{vcftools.meanDP}
\alias{vcftools.mac}
\alias{vcftools.monomorphic}
\alias{vcftools.excessReads}
\alias{vcftools.removeIndels}
\alias{vcftools.maxMiss}
\alias{vcftools.rmLoci}
\title{vcftools Locus Filters}
\usage{
vcftools.minDP(vcf.fn, res.fn, minDP)

vcftools.minQ(vcf.fn, res.fn, minQ)

vcftools.meanDP(vcf.fn, res.fn, meanDP)

vcftools.mac(vcf.fn, res.fn, mac)

vcftools.monomorphic(vcf.fn, res.fn)

vcftools.excessReads(vcf.fn, res.fn, max.reads)

vcftools.removeIndels(vcf.fn, res.fn)

vcftools.maxMiss(vcf.fn, res.fn, max.miss)

vcftools.rmLoci(vcf.fn, res.fn, loci.file.name)
}
\arguments{
\item{vcf.fn}{character string with the relative path and name of the vcf file to be filtered}

\item{res.fn}{character string with the relative path and name of the output vcf file}

\item{minDP}{minimum read depth below which a genotype should be removed}

\item{minQ}{minimum quality below which genotype should be removed}

\item{meanDP}{minimum mean read depth below which loci should be removed}

\item{mac}{minimum minor allele count below which the minor allele should not be called}

\item{max.reads}{maximum number of reads above which a locus is removed}

\item{max.miss}{maximum number of missing reads per individual below which
loci should be removed}

\item{loci.file.name}{character string naming a file with names of loci to remove}
}
\value{
two character strings stating 1) the number of individuals filtered
    out and 2) the number of loci filtered out
}
\description{
These functions filter out loci or genotypes based on the arguments
    provided. Input and results file names should include the path, but no
    extension. Functions return the numbers of individuals and loci kept after
    filtering.
}
