% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcftools.ind.filters.R
\name{vcftools.removeInd}
\alias{vcftools.removeInd}
\title{vcftools Individual Filters}
\usage{
vcftools.removeInd(vcf.fn, res.fn, ind.file.name)
}
\arguments{
\item{vcf.fn}{character string with the relative path and name of the vcf file to be filtered}

\item{res.fn}{character string with the relative path and name of the output vcf file}

\item{ind.file.name}{name of file listing individuals to be removed}
}
\value{
two character strings stating 1) the number of individuals filtered
    out and 2) the number of loci filtered out
}
\description{
These functions filter out individuals based on the arguments
    provided. Input and results file names should include the path, but no
    extension. Functions return the numbers of individuals and loci kept after
    filtering.
}
